Explicitly set coveralls service name to 'github'
Via GitHub Actions, the Coveralls API appears to respond with a 422
'Unprocessable Entity' error unless the service name is set to
'github'. On the other hand, the latest release of
https://pypi.org/project/coveralls apparently defaults to setting the
service name to 'github-actions'. Fortunately, we can override this
default via the COVERALLS_SERVICE_NAME environment variable and resolve
the conflict.

See TheKevJames/coveralls-python#252 for more information.