Enable creation of contextless Invoke tasks
Adds a new `task` utility function (and some helper functions) to the
tasks module. This function allows for the creation of an Invoke task
without the requirement that the decorated function must have an
initial context parameter.

Eventually, this functionality may be provided directly by Invoke, at
which point this code may be replaced by the Invoke equivalent. See
pyinvoke/invoke#445 and pyinvoke/invoke#473 for more information.