Automatically generate tool documentation in tasks (#88)
Replaces the hard-coded tool list descriptions in the docstrings of
both the `format` and `check` tasks with dynamic, automatically
generated lists. 

Previously, each time an additional tool was added to either task,
the docstring for that task would have to be updated manually. Being
rather minor, this change was often forgotten, which resulted in
outdated documentation (see PRs #80, #86, and #87, for example).
These changes resolve that issue by making this documentation
dynamically generate based on the actual tool list run by each task.
In other words, it will no longer be necessary to remember to update
this particular documentation because it now stays up-to-date
automatically.